+++ src/dbg.ts
@@ -1,10 +1,18 @@
 import * as dom from './dom';
 import { logs, TaggedLogger } from './log';
+import * as config from './config';
 
 const log = new TaggedLogger('dbg');
 const { $ } = dom;
 
 export function init() {
+  log.i('Debug mode?', config.DEBUG);
+
+  if (!config.DEBUG)
+    return;
+
+  document.body.classList.add(dom.CSS_DEBUG);
+
   $(dom.ID_RESET_LS).addEventListener('click', () => {
     log.i('#reset-logs:click');
     localStorage.clear();
